Workflow for downsampling chIP-seq
This workflow takes a raw chIP-seq file, aligns it with Bowtie2, selects reads aligning to a list of user input regions, and converts the aligned reads to Fastq format for realignment. The final file will be downsampled to the user input regions and the small file size is ideal for workshop/training situations.